moved components in route specif dirs
