+++ src/routes/Account/index.jsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
-import LogoutButton from '@/components/account/LogoutButton';
+import LogoutButton from './LogoutButton';
 import { useAuth } from '@/hooks/useAuth';
 import styles from './styles.module.css';
