+++ src/routes/Dashboard/index.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import { useScheduleQuery } from '@/hooks/useScheduleQuery';
-import StatusColumn from './components/StatusColumn';
-import StatusTab from './components/StatusTab';
+import StatusColumn from './StatusColumn';
+import StatusTab from './StatusTab';
 import { formatDateFull, sortAndFormatApptByStartTime } from '@/utils/date';
 import { filterByApptStatus } from '@/utils/apptStatus';
 import styles from './styles.module.css';
